pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lightbench"
version = "0.1.0"
dependencies = [
    "python-dotenv",
    "transformers",
    "bitsandbytes",
    "accelerate>=0.26.0",
    "nvidia-ml-py",
    "openai",
    "mistralai",
    "anthropic",
    "pynvml",
]

[project.optional-dependencies]
dev = ["pytest", "sphinx", "myst-parser", "sphinx-book-theme", "sphinx-design"]
api = ["fastapi", "uvicorn"]
notebook = ["ipywidgets"]

[tool.setuptools]
packages = { find = {} }